requirements.txt
requests
python-dotenv
pandas
numpy
scipy
scikit-learn
tqdm
fastapi
uvicorn[standard]
